AudioViewModel rework
