+++ app/src/main/java/com/marat/hvatit/playlistmaker2/presentation/audioplayer/AudioViewModel.kt
@@ -1,5 +1,8 @@
 package com.marat.hvatit.playlistmaker2.presentation.audioplayer
 
+import android.os.Handler
+import android.os.Looper
+import android.util.Log
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
@@ -11,35 +14,54 @@ import com.marat.hvatit.playlistmaker2.domain.api.AudioPlayerInteractor
 class AudioViewModel(private val interactor: AudioPlayerInteractor) : ViewModel() {
 
     private var playerState: MediaPlayerState = MediaPlayerState.Default
-    private val playingState : MediaPlayerState.Playing = MediaPlayerState.Playing(interactor.updateTimer())
+    private var currentState: MediaPlayerState? = null
     private var loadingLiveData = MutableLiveData(playerState)
+
+    private val handler = Handler(Looper.getMainLooper())
+    private val timerRunnable: Runnable = Runnable { updateTimer() }
 
 
     fun getLoadingLiveData(): LiveData<MediaPlayerState> = loadingLiveData
 
-    fun playbackControl(){
-        loadingLiveData.postValue(interactor.playbackControl())
-        //playerState = interactor.playbackControl()
+    fun playbackControl() {
+        currentState = interactor.playbackControl()
+        loadingLiveData.postValue(currentState)
+        if (currentState is MediaPlayerState.Playing) {
+            updateTimer()
+        } else {
+            stopTimer()
+        }
     }
 
-    fun refreshTime(){
-        //playingState.copy(interactor.updateTimer())
-        loadingLiveData.postValue(playingState)
-
-        //loadingLiveData.postValue(playingState.copy(interactor.updateTimer()))
-        //loadingLiveData.postValue(playingState)
-        //loadingLiveData.postValue(MediaPlayerState.Playing(this.interactor.updateTimer()))
+    fun refreshTime() {
+        loadingLiveData.postValue(MediaPlayerState.Playing(interactor.updateTimer()))
     }
 
     fun stopPlayer() {
+        stopTimer()
         interactor.stopPlayer()
         loadingLiveData.postValue(MediaPlayerState.Paused)
     }
 
     fun destroyPlayer() {
+        stopTimer()
         interactor.destroyPlayer()
     }
 
+    fun trackIsDone(){
+        stopTimer()
+        loadingLiveData.postValue(MediaPlayerState.Paused)
+    }
+
+    private fun updateTimer() {
+        Log.e("MediaState", "updateTimer")
+        refreshTime()
+        handler.postDelayed(timerRunnable, 1000L)
+    }
+
+    private fun stopTimer() {
+        handler.removeCallbacks(timerRunnable)
+    }
     companion object {
         fun getViewModelFactory(interactor: AudioPlayerInteractor): ViewModelProvider.Factory =
             viewModelFactory {
